2nd pass
Refactored, but still surely implementing incorrectly. the extension crashes after a knot is made (probably when trying to render an editable instance of <KnotInput>)
