+++ src/contents/all/components/knot/KnotInput/KnotInput.tsx
@@ -2,7 +2,7 @@ import React, { useState, useRef, useEffect } from 'react';
 import { Timeline } from "antd";
 import RichTextEditor from 'react-rte/lib/RichTextEditor';
 
-import { User, Knot } from '../../../global/types';
+import { User, Knot, Log } from '../../../global/types';
 
 import KeyboardShortcut from '../../common/KeyboardShortcut/KeyboardShortcut'
 
@@ -10,28 +10,29 @@ import "../Knot/Knot.scss"
 import "./KnotInput.scss";
 
 interface Props {
+  knot?: string
   createKnot: (bodyHtml: string, bodyText: string) => void
-  deleteKnot: any
-  editKnot: any
-  hasKnots: boolean
+  editKnot?: (bodyHtml: string, bodyText: string) => void
+  hasKnots?: boolean
   show: boolean
 }
 
-const KnotInput = ({createKnot, hasKnots, show}: Props) => {
+const KnotInput = ({knot, createKnot, editKnot, hasKnots, show}: Props) => {
 
   const knotEditor = useRef(null)
 
   const [submit, setSubmit] = useState(false)
   const [showFooter, setShowFooter] = useState(false)
   const [body, setBody] = useState(RichTextEditor.createEmptyValue())
-  const [knot, setKnot] = useState<Knot>(null)
 
   const [loading, setLoading] = useState(false)
-  const [deleteLoading, setDeleteLoading] = useState(false)
 
   useEffect(() => {
     const targetIsRef = knotEditor.hasOwnProperty("current")
     const currentTarget = targetIsRef ? knotEditor.current : knotEditor;
+    if (knot) {
+      setBody(knot)
+    }
     if (currentTarget)
       currentTarget.addEventListener("keydown", onKeyDown)
     return () => {
@@ -54,10 +55,18 @@ const KnotInput = ({createKnot, hasKnots, show}: Props) => {
     const tempText = value.getEditorState().getCurrentContent().getPlainText() // value.toString("markdown")
 
     if (submit) {
-      createKnot(
-        tempHtml.substring(0, tempHtml.length-11), // remove last line <p><br></p>
-        tempText.replace(/(\r\n|\n|\r)/gm, ""), // remove trailing line break
-      )
+      if (knot) {
+        editKnot(
+          tempHtml.substring(0, tempHtml.length-11), // remove last line <p><br></p>
+          tempText.replace(/(\r\n|\n|\r)/gm, ""), // remove trailing line break
+        )
+      }
+      else {
+        createKnot(
+          tempHtml.substring(0, tempHtml.length-11), // remove last line <p><br></p>
+          tempText.replace(/(\r\n|\n|\r)/gm, ""), // remove trailing line break
+        )
+      }
       setSubmit(false)
       setBody(RichTextEditor.createEmptyValue())
     }
@@ -71,10 +80,18 @@ const KnotInput = ({createKnot, hasKnots, show}: Props) => {
   }
 
   const onSubmit = () => {
-    createKnot(
-      body.toString("html"),
-      body.getEditorState().getCurrentContent().getPlainText()
-    )
+    if (knot) {
+      editKnot(
+        body.toString("html"),
+        body.getEditorState().getCurrentContent().getPlainText()
+      )
+    }
+    else {
+      createKnot(
+        body.toString("html"),
+        body.getEditorState().getCurrentContent().getPlainText()
+      )
+    }
     setSubmit(false)
     setBody(RichTextEditor.createEmptyValue())
   }
